refactor(studio-pack): extract install logic into useEnsureStudioPack
